Animation.java - Removed the cellWidth and cellHeight parameters.
GameView.java - Added a new int named currentHP which will keep track of the player's current hp. Added 3 new booleans isJumping, movingLeft, and movingRight. added an int for gravity, to have the character always move down, which eliminates the holding of the jump action. Added an if statement around the item spawn to keep the spawns low as well as added effects to the items, coffee now heals the player 100 hp, and pills now reduce the player's hp by 100. Edited the onTouch Event. Made the if statements smaller and a little easier to understand. added the booleans for moving and jumping in the on touch events. Added velocity in the updateLogic function. Added the item collision in the updateCollision function.

Items.java - Deleted.

Sprite.java - started to make a hit box function in the Sprite.java, made a few privates more accessible.